Optimized as many things as possible
